Moving files back to private section
